NE import names problem

+++ plugins/SunFlower.Ne/Headers/NeModule.cs
@@ -5,5 +5,5 @@ namespace SunFlower.Ne.Headers;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct NeModule
 {
-    public UInt32 ImportOffset;
+    public UInt16 ImportOffset;
 }
